Basic Navigation Bar Finished
The code for my navigation bar is basically finished:  I have the list of links, I have my background-color and link colors set so they contrast.

I am going to create blank pages and copy-and-paste the navigation bar code onto them so the navigation loads the same on each page.

The remaining things to tweak may be rounding the corners and adjusting the colors if necessary.