Added Placeholder Files for About Me and Science Pages
These are blank pages with the navigation bar code copy-and-pasted so that it is uniform throughout the site.

Since my links are technically loading different .html files each time and not #linking to an anchor within the same page I am using this approach.  There may be a better way to do it but it works.